irc: do joins in case there is no motd
To suit ircd refusing joins before MOTD ended, the startup joins are
done once the RPL_ENDOFMOTD is received. In some cases, an ircd
configuration can be messed up, and contains no MOTD, this return code
won't be received, and joins won't be done.

To fix that, simply add a case doing the joins on ERR_NOMOTD.

Signed-off-by: bleader <bleader@ratonland.org>